Updates for Unit Decimal points
To address the use case where a queryString includes the unitdp=4 param we need to change objects with attributes that support more than two decimal points to to change the number datatype from double to float.

Added missing unitdp to methods ...

Added to the following ..

BankTransactions
- getBankTransaction
- createBankTransactions
- updateBankTransaction
- updateOrCreate

CreditNotes
- getCreditNotes
- createCreditNotes
- updateCreditNote
- updateOrCreateCreditNotes

Invoices
- getInvoice
- createInvoices
- updateInvoice
- updateOrCreateInvoices

Items
- getItem
- createItems
- updateItem
- updateOrCreateItems

Receipts
- getReceipts
- getReceipt
- createReceipt
- updateReceipt

Objects that use LineItems are impacted
- BankTransations
- CreditNotes
- Invoices
- Receipts

the two properties that changed from double to float
- Quantity
- UnitAmount

The Purchase object property changed from double to float
- UnitPrice

The Item object has the property changed from double to float
- QuantityOnHand

Added ARCHIVED Status to Manual Journal object
Added SalesTaxPeriod NONE to organisation object
Added Reference attribute to Prepayment object
Added VOIDED status to Receipt object
Added DELETED status to RepeatingInvoice object

Updated corresponding Unit Tests

+++ src/main/java/com/xero/models/accounting/LineItem.java
@@ -44,11 +44,11 @@ public class LineItem {
 
   
   @JsonProperty("Quantity")
-  private Double quantity;
+  private Float quantity;
 
   
   @JsonProperty("UnitAmount")
-  private Double unitAmount;
+  private Float unitAmount;
 
   
   @JsonProperty("ItemCode")
@@ -121,7 +121,7 @@ public class LineItem {
     this.description = description;
   }
 
-  public LineItem quantity(Double quantity) {
+  public LineItem quantity(Float quantity) {
     this.quantity = quantity;
     return this;
   }
@@ -131,15 +131,15 @@ public class LineItem {
    * @return quantity
   **/
   @ApiModelProperty(value = "LineItem Quantity")
-  public Double getQuantity() {
+  public Float getQuantity() {
     return quantity;
   }
 
-  public void setQuantity(Double quantity) {
+  public void setQuantity(Float quantity) {
     this.quantity = quantity;
   }
 
-  public LineItem unitAmount(Double unitAmount) {
+  public LineItem unitAmount(Float unitAmount) {
     this.unitAmount = unitAmount;
     return this;
   }
@@ -149,11 +149,11 @@ public class LineItem {
    * @return unitAmount
   **/
   @ApiModelProperty(value = "LineItem Unit Amount")
-  public Double getUnitAmount() {
+  public Float getUnitAmount() {
     return unitAmount;
   }
 
-  public void setUnitAmount(Double unitAmount) {
+  public void setUnitAmount(Float unitAmount) {
     this.unitAmount = unitAmount;
   }
